producer/requirements.txt
kafka-python==2.0.2
pandas==2.1.4
pyarrow
fastparquet
